PYIC-7304: improve styling of events w/ journeyContext
